requirements.txt
absl-py==0.7.0
astor==0.7.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cloudpickle==0.7.0
cycler==0.10.0
dask==1.1.0
decorator==4.3.2
Flask==1.0.2
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
idna==2.8
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
nbformat==4.4.0
networkx==2.2
numpy==1.16.0
pandas==0.23.4
Pillow==5.4.1
pkg-resources==0.0.0
plotly==3.5.0
protobuf==3.6.1
pyparsing==2.3.1
python-dateutil==2.7.5
pytz==2018.9
PyWavelets==1.0.1
requests==2.21.0
retrying==1.3.3
scikit-image==0.14.2
scipy==1.2.0
six==1.12.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
toolz==0.9.0
traitlets==4.3.2
urllib3==1.24.1
Werkzeug==0.14.1
